Corrections To Basic Running
Found out that an unlocked value of "" errored out the program because it couldn't be parsed as a number in the Mission constructor.

The UI portion of the program shouldn't output the ID of the mission being completed.  That can be done in the console, where additional layers of information are welcome.